- Split models.py out into separate files - Add some docstrings to models
