+++ wagtailmenus/models/__init__.py
@@ -1,0 +1,10 @@
+from .pages import MenuPage
+from .menus import (
+    Menu, MenuFromRootPage, MenuWithMenuItems,
+    AbstractMainMenu, AbstractFlatMenu,
+    MainMenu, FlatMenu
+)
+from .menuitems import (
+    MenuItem, AbstractMenuItem, AbstractMainMenuItem, AbstractFlatMenuItem,
+    MainMenuItem, FlatMenuItem
+)
