+++ wagtailmenus/models/menus.py
@@ -1,209 +1,31 @@
 from __future__ import absolute_import, unicode_literals
 
 from collections import defaultdict
-from copy import copy
 
 from django.db import models
-from django.core.exceptions import ValidationError
+from django.core.exceptions import ImproperlyConfigured, ValidationError
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.functional import cached_property
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 
 from modelcluster.models import ClusterableModel
-from modelcluster.fields import ParentalKey
 
 from wagtail.wagtailadmin.edit_handlers import (
-    FieldPanel, PageChooserPanel, MultiFieldPanel, InlinePanel)
-from wagtail.wagtailcore.models import Page, Orderable
+    FieldPanel, MultiFieldPanel, InlinePanel)
+from wagtail.wagtailcore.models import Page
 
 from wagtailmenus import app_settings
-from .forms import FlatMenuAdminForm
-from .managers import MenuItemManager
-from .panels import menupage_settings_panels
+from ..forms import FlatMenuAdminForm
 
 
-class MenuPage(Page):
-    repeat_in_subnav = models.BooleanField(
-        verbose_name=_("repeat in sub-navigation"),
-        default=False,
-        help_text=_(
-            "If checked, a link to this page will be repeated alongside it's "
-            "direct children when displaying a sub-navigation for this page."
-        ),
-    )
-    repeated_item_text = models.CharField(
-        verbose_name=_('repeated item link text'),
-        max_length=255,
-        blank=True,
-        help_text=_(
-            "e.g. 'Section home' or 'Overview'. If left blank, the page title "
-            "will be used."
-        )
-    )
-
-    settings_panels = menupage_settings_panels
-
-    class Meta:
-        abstract = True
-
-    def modify_submenu_items(
-        self, menu_items, current_page, current_ancestor_ids, current_site,
-        allow_repeating_parents, apply_active_classes, original_menu_tag,
-        menu_instance
-    ):
-        """
-        Make any necessary modifications to `menu_items` and return the list
-        back to the calling menu tag to render in templates. Any additional
-        items added should have a `text` and `href` attribute as a minimum.
-
-        `original_menu_tag` should be one of 'main_menu', 'section_menu' or
-        'children_menu', which should be useful when extending/overriding.
-        """
-        if (allow_repeating_parents and menu_items and self.repeat_in_subnav):
-            """
-            This page should have a version of itself repeated alongside
-            children in the subnav, so we create a new item and prepend it to
-            menu_items.
-            """
-            menu_items.insert(0, self.get_repeated_menu_item(
-                current_page, current_site, apply_active_classes,
-                original_menu_tag
-            ))
-        return menu_items
-
-    def has_submenu_items(
-        self, current_page, allow_repeating_parents, original_menu_tag,
-        menu_instance
-    ):
-        """
-        When rendering pages in a menu template a `has_children_in_menu`
-        attribute is added to each page, letting template developers know
-        whether or not the item has a submenu that must be rendered.
-
-        By default, we return a boolean indicating whether the page has
-        suitable child pages to include in such a menu. But, if you are
-        overriding the `modify_submenu_items` method to programatically add
-        items that aren't child pages, you'll likely need to alter this method
-        too, so the template knows there are sub items to be rendered.
-        """
-        return menu_instance.page_has_children(self)
-
-    def get_repeated_menu_item(
-        self, current_page, current_site, apply_active_classes,
-        original_menu_tag
-    ):
-        """Return something that can be used to display a 'repeated' menu item
-        for this specific page."""
-        menuitem = copy(self)
-        setattr(menuitem, 'text', self.repeated_item_text or self.title)
-        setattr(menuitem, 'href', self.relative_url(current_site))
-        active_class = ''
-        if apply_active_classes and self == current_page:
-            active_class = app_settings.ACTIVE_CLASS
-        setattr(menuitem, 'active_class', active_class)
-        return menuitem
-
-
-@python_2_unicode_compatible
-class MenuItem(models.Model):
-    allow_subnav = False
-
-    link_page = models.ForeignKey(
-        'wagtailcore.Page',
-        verbose_name=_('link to an internal page'),
-        blank=True,
-        null=True,
-        on_delete=models.CASCADE,
-    )
-    link_url = models.CharField(
-        verbose_name=_('link to a custom URL'),
-        max_length=255,
-        blank=True,
-        null=True,
-    )
-    link_text = models.CharField(
-        verbose_name=_('link text'),
-        max_length=255,
-        blank=True,
-        help_text=_(
-            "Provide the text to use for a custom URL, or set on an internal "
-            "page link to use instead of the page's title."
-        ),
-    )
-    handle = models.CharField(
-        verbose_name=_('handle'),
-        max_length=100,
-        blank=True,
-        help_text=_(
-            "Use this field to optionally specify an additional value for "
-            "each menu item, which you can then reference in custom menu "
-            "templates."
-        )
-    )
-    url_append = models.CharField(
-        verbose_name=_("append to URL"),
-        max_length=255,
-        blank=True,
-        help_text=_(
-            "Use this to optionally append a #hash or querystring to the "
-            "above page's URL."
-        )
-    )
-
-    objects = MenuItemManager()
-
-    class Meta:
-        abstract = True
-        verbose_name = _("menu item")
-        verbose_name_plural = _("menu items")
-
-    @property
-    def menu_text(self):
-        return self.link_text or self.link_page.title
-
-    def relative_url(self, site=None):
-        if self.link_page:
-            return self.link_page.relative_url(site) + self.url_append
-        return self.link_url + self.url_append
-
-    def clean(self, *args, **kwargs):
-        super(MenuItem, self).clean(*args, **kwargs)
-
-        if self.link_url and not self.link_text:
-            raise ValidationError({
-                'link_text': [
-                    _("This must be set if you're linking to a custom URL."),
-                ]
-            })
-
-        if not self.link_url and not self.link_page:
-            raise ValidationError({
-                'link_url': [
-                    _("This must be set if you're not linking to a page."),
-                ]
-            })
-
-        if self.link_url and self.link_page:
-            raise ValidationError(_(
-                "You cannot link to both a page and URL. Please review your "
-                "link and clear any unwanted values."
-            ))
-
-    def __str__(self):
-        return self.menu_text
-
-    panels = (
-        PageChooserPanel('link_page'),
-        FieldPanel('link_url'),
-        FieldPanel('url_append'),
-        FieldPanel('link_text'),
-        FieldPanel('handle'),
-        FieldPanel('allow_subnav'),
-    )
-
+# ########################################################
+# Base
+# ########################################################
 
 class Menu(object):
+    """A base class that all other 'menu' classes should inherit from."""
+
     max_levels = 1
     use_specific = app_settings.USE_SPECIFIC_AUTO
     pages_for_display = None
@@ -247,10 +69,8 @@ class Menu(object):
 
     @cached_property
     def page_children_dict(self):
-        """
-        Returns a dictionary of lists, where the keys are 'path' values for
-        pages, and the value is a list of children pages for that page.
-        """
+        """Returns a dictionary of lists, where the keys are 'path' values for
+        pages, and the value is a list of children pages for that page."""
         children_dict = defaultdict(list)
         for page in self.pages_for_display:
             children_dict[page.path[:-page.steplen]].append(page)
@@ -261,14 +81,15 @@ class Menu(object):
         return self.page_children_dict.get(page.path, [])
 
     def page_has_children(self, page):
-        """
-        Return a boolean indicating whether a given page has any relevant
-        child pages.
-        """
+        """Return a boolean indicating whether a given page has any relevant
+        child pages."""
         return page.path in self.page_children_dict
 
 
 class MenuFromRootPage(Menu):
+    """A 'menu' that is instantiated with a 'root page', and whose 'menu items'
+    consist solely of ancestors of that page."""
+
     root_page = None
 
     def __init__(self, root_page, max_levels, use_specific):
@@ -279,10 +100,8 @@ class MenuFromRootPage(Menu):
 
     @cached_property
     def pages_for_display(self):
-        """
-        Returns a list of pages for rendering all levels of the menu. All pages
-        must be live, not expired, and set to show in menus.
-        """
+        """Returns a list of pages for rendering all levels of the menu. All
+        pages must be live, not expired, and set to show in menus."""
         pages = Page.objects.filter(
             depth__gt=self.root_page.depth,
             depth__lte=self.root_page.depth + self.max_levels,
@@ -308,17 +127,23 @@ class MenuFromRootPage(Menu):
 
 
 class MenuWithMenuItems(ClusterableModel, Menu):
+    """A base model class for menus who's 'menu_items' are defined by
+    a set of 'menu item' model instances."""
 
     class Meta:
         abstract = True
 
+    def get_menu_items_manager(self):
+        raise NotImplementedError(
+            'MenuWithMenuItems subclasses must define their own '
+            'get_menu_items_manager() method.'
+        )
+
     @cached_property
     def top_level_items(self):
-        """
-        Return a list of menu_items with link_page objects supplemented with
-        'specific' pages where appropriate.
-        """
-        items_qs = self.menu_items.for_display()
+        """Return a list of menu items with link_page objects supplemented with
+        'specific' pages where appropriate."""
+        items_qs = self.get_menu_items_manager().for_display()
         if self.use_specific < app_settings.USE_SPECIFIC_TOP_LEVEL:
             return items_qs.all()
 
@@ -336,11 +161,9 @@ class MenuWithMenuItems(ClusterableModel, Menu):
 
     @cached_property
     def pages_for_display(self):
-        """
-        Returns a list of pages for rendering the entire menu (excluding those
-        chosen as menu items). All pages must be live, not expired, and set to
-        show in menus.
-        """
+        """Return a list of pages for rendering the entire menu (excluding
+        those chosen as menu items). All pages must be live, not expired, and
+        set to show in menus."""
 
         # Build a queryset to get pages for all levels
         all_pages = Page.objects.none()
@@ -376,6 +199,11 @@ class MenuWithMenuItems(ClusterableModel, Menu):
         return all_pages
 
 
+# ########################################################
+# Abstract
+# ########################################################
+
+@python_2_unicode_compatible
 class AbstractMainMenu(MenuWithMenuItems):
     site = models.OneToOneField(
         'wagtailcore.Site',
@@ -415,14 +243,27 @@ class AbstractMainMenu(MenuWithMenuItems):
 
     @classmethod
     def get_for_site(cls, site):
-        """
-        Get a mainmenu instance for the site.
-        """
+        """Get a mainmenu instance for the site."""
         instance, created = cls.objects.get_or_create(site=site)
         return instance
 
     def __str__(self):
         return _('Main menu for %s') % (self.site.site_name or self.site)
+
+    def get_menu_items_manager(self):
+        try:
+            return getattr(self, app_settings.MAIN_MENU_ITEMS_RELATED_NAME)
+        except AttributeError:
+            raise ImproperlyConfigured(
+                "'%s' isn't a valid relationship name for accessing menu "
+                "items from %s. Check that your "
+                "`WAGTAILMENUS_MAIN_MENU_ITEMS_RELATED_NAME` setting matches "
+                "the `related_name` used on your MenuItem model's "
+                "`ParentalKey` field." % (
+                    app_settings.MAIN_MENU_ITEMS_RELATED_NAME,
+                    self.__class__.__name__
+                )
+            )
 
     panels = (
         InlinePanel(
@@ -436,28 +277,7 @@ class AbstractMainMenu(MenuWithMenuItems):
     )
 
 
-class MainMenu(AbstractMainMenu):
-    pass
-
-
-class AbstractMainMenuItem(Orderable, MenuItem):
-    allow_subnav = models.BooleanField(
-        verbose_name=_("allow sub-menu for this item"),
-        default=True,
-        help_text=_(
-            "NOTE: The sub-menu might not be displayed, even if checked. "
-            "It depends on how the menu is used in this project's templates."
-        )
-    )
-
-    class Meta:
-        abstract = True
-
-
-class MainMenuItem(AbstractMainMenuItem):
-    menu = ParentalKey('MainMenu', related_name="menu_items")
-
-
+@python_2_unicode_compatible
 class AbstractFlatMenu(MenuWithMenuItems):
     site = models.ForeignKey(
         'wagtailcore.Site',
@@ -519,10 +339,8 @@ class AbstractFlatMenu(MenuWithMenuItems):
     @classmethod
     def get_for_site(cls, handle, site,
                      fall_back_to_default_site_menus=False):
-        """
-        Get a FlatMenu instance with a matching `handle` for the `site`
-        provided - or for the 'default' site if not found.
-        """
+        """Get a FlatMenu instance with a matching `handle` for the `site`
+        provided - or for the 'default' site if not found."""
         menu = cls.objects.filter(handle__exact=handle, site=site).first()
         if(
             menu is None and fall_back_to_default_site_menus and
@@ -536,9 +354,25 @@ class AbstractFlatMenu(MenuWithMenuItems):
     def __str__(self):
         return '%s (%s)' % (self.title, self.handle)
 
+    def get_menu_items_manager(self):
+        try:
+            return getattr(self, app_settings.FLAT_MENU_ITEMS_RELATED_NAME)
+        except AttributeError:
+            raise ImproperlyConfigured(
+                "'%s' isn't a valid relationship name for accessing menu "
+                "items from %s. Check that your "
+                "`WAGTAILMENUS_FLAT_MENU_ITEMS_RELATED_NAME` setting matches "
+                "the `related_name` used on your MenuItem model's "
+                "`ParentalKey` field." % (
+                    app_settings.FLAT_MENU_ITEMS_RELATED_NAME,
+                    self.__class__.__name__
+                )
+            )
+
     def clean(self, *args, **kwargs):
-        # Raise validation error for unique_together constraint, as it's not
-        # currently handled properly by wagtail
+        """Raise validation error for unique_together constraint, as it's not
+        currently handled properly by wagtail."""
+
         clashes = self.__class__.objects.filter(site=self.site,
                                                 handle=self.handle)
         if self.pk:
@@ -573,23 +407,15 @@ class AbstractFlatMenu(MenuWithMenuItems):
     )
 
 
-class FlatMenu(AbstractFlatMenu):
+# ########################################################
+# Concrete
+# ########################################################
+
+class MainMenu(AbstractMainMenu):
+    """The default model for 'main menu' instances."""
     pass
 
 
-class AbstractFlatMenuItem(Orderable, MenuItem):
-    allow_subnav = models.BooleanField(
-        verbose_name=_("allow sub-menu for this item"),
-        default=False,
-        help_text=_(
-            "NOTE: The sub-menu might not be displayed, even if checked. "
-            "It depends on how the menu is used in this project's templates."
-        )
-    )
-
-    class Meta:
-        abstract = True
-
-
-class FlatMenuItem(AbstractFlatMenuItem):
-    menu = ParentalKey('FlatMenu', related_name="menu_items")
+class FlatMenu(AbstractFlatMenu):
+    """The default model for 'flat menu' instances."""
+    pass
